Use fixed variant for top app bar on my account
- move nav-menu.component into app.component
