+++ src/SIL.XForge.Scripture/ClientApp/src/app/app.component.scss
@@ -10,8 +10,27 @@
     z-index: 10;
   }
 
+  > mdc-drawer {
+    > mdc-drawer-header {
+      padding: 0;
+
+      > .project-select {
+        width: 100%;
+      }
+    }
+
+    .mdc-list-item {
+      cursor: pointer;
+    }
+  }
+
   > header {
     position: absolute;
+
+    &.overlay-drawer {
+      z-index: 7;
+    }
+
     mdc-top-app-bar-title {
       cursor: pointer;
     }
@@ -20,6 +39,18 @@
       mdc-menu {
         min-width: 200px;
       }
+    }
+
+    .mdc-top-app-bar--short-collapsed {
+      width: 255px !important;
+    }
+
+    .mdc-top-app-bar__row {
+      height: 56px;
+    }
+
+    .mdc-top-app-bar__section {
+      padding: 4px;
     }
   }
 
@@ -34,9 +65,5 @@
       flex: 1;
       padding: 14px 20px;
     }
-  }
-
-  .mdc-top-app-bar--short-collapsed {
-    width: 255px !important;
   }
 }
