--- src/SIL.XForge.Scripture/ClientApp/src/app/nav-menu/nav-menu.component.html
@@ -1,117 +1,0 @@
-<mdc-drawer
-  id="menu-drawer"
-  *ngIf="selectedProject != null"
-  [drawer]="isDrawerPermanent ? 'permanent' : 'modal'"
-  [open]="isExpanded"
-  (closed)="drawerCollapsed()"
->
-  <mdc-drawer-header>
-    <div *ngIf="isDrawerPermanent" class="top-app-bar-adjust-height"></div>
-    <mdc-select *ngIf="projects != null" class="project-select" (valueChange)="projectChanged($event.value)">
-      <mdc-menu>
-        <mdc-list>
-          <mdc-list-item *ngFor="let project of projects" [value]="project.id">{{ project.projectName }}</mdc-list-item>
-          <mdc-list-divider></mdc-list-divider>
-          <mdc-list-item value="*connect-project*">
-            <mdc-icon mdcListItemGraphic>add</mdc-icon>
-            Connect project
-          </mdc-list-item>
-        </mdc-list>
-      </mdc-menu>
-    </mdc-select>
-  </mdc-drawer-header>
-  <mdc-drawer-content>
-    <mdc-list id="menu-list">
-      <mdc-list-item
-        *ngIf="selectedProject.translateConfig?.enabled"
-        (selectionChange)="translateVisible = !translateVisible"
-      >
-        <mdc-icon mdcListItemGraphic>translate</mdc-icon>
-        Translate
-        <mdc-icon mdcListItemMeta>keyboard_arrow_{{ translateVisible ? "down" : "right" }}</mdc-icon>
-      </mdc-list-item>
-      <div *ngIf="translateVisible">
-        <mdc-list-item
-          [routerLink]="['/projects', selectedProject.id, 'translate']"
-          [activated]="rlaTranslateOverview.isActive"
-          routerLinkActive
-          [routerLinkActiveOptions]="{ exact: true }"
-          #rlaTranslateOverview="routerLinkActive"
-          (selectionChange)="itemSelected()"
-        >
-          <mdc-icon mdcListItemGraphic>apps</mdc-icon>
-          Overview
-        </mdc-list-item>
-        <mdc-list-item
-          *ngFor="let text of texts"
-          [routerLink]="['/projects', selectedProject.id, 'translate', text.id]"
-          [activated]="rla.isActive"
-          routerLinkActive
-          #rla="routerLinkActive"
-          (selectionChange)="itemSelected()"
-        >
-          <mdc-icon mdcListItemGraphic>book</mdc-icon>
-          {{ text.name }}
-        </mdc-list-item>
-      </div>
-      <mdc-list-divider *ngIf="selectedProject.translateConfig?.enabled"></mdc-list-divider>
-      <mdc-list-item
-        *ngIf="selectedProject.checkingConfig?.enabled"
-        (selectionChange)="checkingVisible = !checkingVisible"
-      >
-        <mdc-icon mdcListItemGraphic>question_answer</mdc-icon>
-        Community Checking
-        <mdc-icon mdcListItemMeta>keyboard_arrow_{{ checkingVisible ? "down" : "right" }}</mdc-icon>
-      </mdc-list-item>
-      <div *ngIf="checkingVisible">
-        <mdc-list-item
-          [routerLink]="['/projects', selectedProject.id, 'checking']"
-          [activated]="rlaCheckingOverview.isActive"
-          routerLinkActive
-          [routerLinkActiveOptions]="{ exact: true }"
-          #rlaCheckingOverview="routerLinkActive"
-          (selectionChange)="itemSelected()"
-        >
-          <mdc-icon mdcListItemGraphic>apps</mdc-icon>
-          Overview
-        </mdc-list-item>
-        <mdc-list-item
-          *ngFor="let text of texts"
-          [routerLink]="['/projects', selectedProject.id, 'checking', text.id]"
-          [activated]="rla.isActive"
-          routerLinkActive
-          #rla="routerLinkActive"
-          (selectionChange)="itemSelected()"
-        >
-          <mdc-icon mdcListItemGraphic>book</mdc-icon>
-          {{ text.name }}
-        </mdc-list-item>
-      </div>
-      <div *ngIf="(isProjectAdmin$ | async)">
-        <mdc-list-divider *ngIf="selectedProject.checkingConfig?.enabled"></mdc-list-divider>
-        <mdc-list-item
-          [routerLink]="['/projects', selectedProject.id, 'sync']"
-          [activated]="rlaSync.isActive"
-          routerLinkActive
-          #rlaSync="routerLinkActive"
-          id="sync-item"
-          (selectionChange)="itemSelected()"
-        >
-          <mdc-icon mdcListItemGraphic>sync</mdc-icon>
-          Synchronize
-        </mdc-list-item>
-        <mdc-list-item
-          [routerLink]="['/projects', selectedProject.id, 'settings']"
-          [activated]="rlaSettings.isActive"
-          routerLinkActive
-          #rlaSettings="routerLinkActive"
-          id="settings-item"
-          (selectionChange)="itemSelected()"
-        >
-          <mdc-icon mdcListItemGraphic>settings</mdc-icon>
-          Settings
-        </mdc-list-item>
-      </div>
-    </mdc-list>
-  </mdc-drawer-content>
-</mdc-drawer>
