--- src/SIL.XForge.Scripture/ClientApp/src/app/nav-menu/nav-menu.component.scss
@@ -1,18 +1,0 @@
-@import '~src/global';
-
-mdc-drawer {
-  > mdc-drawer-header {
-    padding: 0;
-    > .top-app-bar-adjust-height {
-      height: 56px;
-    }
-  }
-}
-
-.project-select {
-  width: 100%;
-}
-
-.mdc-list-item {
-  cursor: pointer;
-}
