+++ src/SIL.XForge.Scripture/ClientApp/src/app/project-deleted-dialog/project-deleted-dialog.component.html
